Log fix for esp idf 5.1

+++ demos/sample_azure_iot_pnp/sample_azure_iot_pnp.c
@@ -361,7 +361,7 @@ static void prvAzureDemoTask( void * pvParameters )
                                            &pulIothubHostnameLength, &pucIotHubDeviceId,
                                            &pulIothubDeviceIdLength ) ) != 0 )
         {
-            LogError( ( "Failed on sample_dps_entry!: error code = 0x%08x\r\n", ulStatus ) );
+            LogError( ( "Failed on sample_dps_entry!: error code = 0x%08x\r\n", ( uint16_t ) ulStatus ) );
             return;
         }
     #endif /* democonfigENABLE_DPS_SAMPLE */
